+++ include/world/SquareObstacle.h
@@ -15,7 +15,7 @@ public:
 
     virtual bool isInArea(int x, int y) const;
     virtual bool isInArea(const Obstacle& obstacle) const;
-    virtual int getDistance(int x, int y, double angle) const;
+    virtual int getDistance(int x, int y, int angle) const;
 
 };
